+++ app/server/api/character/draft/index.delete.ts
@@ -1,0 +1,103 @@
+import { and, eq, isNull } from 'drizzle-orm'
+import { z } from 'zod'
+import { getValidated } from '~/utils/h3'
+import { badRequest, paymentRequired, unauthorized } from '~/utils/nuxt'
+import { isPlanPastDue } from '~/utils/plan'
+import { characterDraftLocalizations, characterDrafts, characters } from '~~/db/schema'
+
+export default eventHandler(async (event) => {
+  const orm = event.context.orm
+  const user = event.context.user
+
+  if (!user)
+    throw unauthorized()
+
+  if (isPlanPastDue(user))
+    throw paymentRequired()
+
+  const { characterId } = await getValidated(event, 'body', z.object({
+    characterId: z.number().optional(),
+  }))
+
+  const existingDraft = await orm.query.characterDrafts.findFirst({
+    where: and(
+      eq(characterDrafts.creatorId, user.id),
+      characterId
+        ? eq(characterDrafts.characterId, characterId)
+        : isNull(characterDrafts.characterId),
+    ),
+    with: {
+      characterDraftLocalizations: {
+        where: eq(characterDraftLocalizations.locale, 'en-US'),
+      },
+      character: !characterId
+        ? undefined
+        : {
+            columns: {
+              categoryId: true,
+            },
+          },
+    },
+  })
+
+  if (!existingDraft) {
+    throw badRequest(characterId
+      ? `You do not have a pending character draft for \`characterId\` ${characterId}`
+      : 'You do not have a pending character draft')
+  }
+
+  if (characterId) {
+    const character = await orm.query.characters.findFirst({
+      where: eq(characters.id, characterId),
+      with: {
+        characterLocalizations: true,
+        usernames: true,
+      },
+    })
+
+    if (!character)
+      throw badRequest(`Character with \`characterId\` ${characterId} not found`)
+
+    const ptLocalization = character.characterLocalizations.find(localization => localization.locale === 'pt-BR')
+    const enLocalization = character.characterLocalizations.find(localization => localization.locale === 'en-US')
+
+    await orm.batch([
+      orm.update(characterDrafts).set({
+        id: existingDraft.id,
+        creatorId: user.id,
+        characterId: character.id,
+        prompt: character.prompt,
+        data: {
+          username: character.usernames!.username,
+          categoryId: character.categoryId,
+        },
+      }),
+      orm.update(characterDraftLocalizations).set({
+        name: ptLocalization?.name,
+        description: ptLocalization?.description,
+        instructions: ptLocalization?.instructions,
+        locale: 'pt-BR',
+        characterDraftId: existingDraft.id,
+      }).where(and(
+        eq(characterDraftLocalizations.locale, 'pt-BR'),
+        eq(characterDraftLocalizations.characterDraftId, existingDraft.id),
+      )),
+      orm.update(characterDraftLocalizations).set({
+        name: enLocalization?.name,
+        description: enLocalization?.description,
+        instructions: enLocalization?.instructions,
+        locale: 'en-US',
+        characterDraftId: existingDraft.id,
+      }).where(and(
+        eq(characterDraftLocalizations.locale, 'en-US'),
+        eq(characterDraftLocalizations.characterDraftId, existingDraft.id),
+      )),
+    ])
+  }
+  else {
+    await orm.batch([
+      orm.delete(characterDrafts).where(eq(characterDrafts.id, existingDraft.id)),
+      orm.delete(characterDraftLocalizations).where(eq(characterDraftLocalizations.characterDraftId, existingDraft.id)),
+    ])
+  }
+})