+++ pm/CHANGELOG.md
@@ -6,6 +6,10 @@ description: A detailed, chronological record of all changes made to a project, 
 
 ## [Unreleased]
 
+### Added
+
+- Added option to delete character draft edition
+
 ## [0.1.0-9] - 2025-05-25
 
 ### Fixed